backend/requirements/base.txt
aioredis
alembic
bcrypt
databases[asyncpg]
orjson
psycopg2
asyncpg
python-jose
SQLAlchemy
httpx
strawberry-graphql[fastapi]
pydantic[email]
fastapi
uvicorn[standard]

sentry-sdk
